refactor, fix: Use on click listener instead of on check changed listener in Local library Manager
Fixes a weird bug that probably happens when the ListView recycles a list item with its CheckBox checked, but the current one having to be unchecked. Code unchecks the CheckBox, but that also triggers the on check changed listener, which then tries to remove that Local library from the enabled Local libraries list. That fails as it's not in the list, which causes an index -1 exception.

Simple fix: Just run the enable/disable logic on clicks only, not check changes.